refactor: move sol builder to own file
